template/{{ project_slug }}/pyproject.toml
[project]
name = "{{ project_slug }}"
version = "{{ project_version }}"
description = "{{ project_description }}"
authors = [
  { name = "{{ project_author_name }}" }
]
readme = "README.md"
requires-python = ">= {{ python_version }}"
license = { text = "{{ project_open_source_license if project_open_source_license != 'No license file' else '' }}" }
dependencies = [
    "numpy",
    "pandas",
    "matplotlib",
    "seaborn",
    "plotly",
    "jupyter",
    "jupyterlab",
    "python-dotenv",
    "missingno",
]

[project.optional-dependencies]
dev = [
    "black",
    "invoke",
    "nbdime",
    "ruff",
    "pytest",
    "sphinx",
    "snakeviz",
]

# ML supervisado: clasificación + PCA opcional como preprocesado
supervisado = [
    "scikit-learn",   # modelos, PCA, pipelines, métricas
    "scipy",          # stats, tests de normalidad, dendrogramas
    "lightgbm",
    "xgboost",
    "statsmodels",
    "imbalanced-learn",
    "joblib",
]

# ML no supervisado: clustering + reducción de dimensionalidad
no_supervisado = [
    "scikit-learn",   # KMeans, AgglomerativeClustering, PCA, DBSCAN
    "scipy",          # dendrogramas y clustering jerárquico
    "yellowbrick",    # SilhouetteVisualizer
    "umap-learn",     # reducción de dimensionalidad no lineal (UMAP)
    "hdbscan",        # clustering basado en densidad jerárquico
    "joblib",
]

# Redes neuronales: deep learning con PyTorch + Polars
redes_neuronales = [
    "torch",
    "torchvision",
    "torchaudio",
    "torchmetrics[visual]",
    "tensorboard",
    "polars",         # carga eficiente de datos con dtypes explícitos
    "scikit-learn",   # PCA, métricas, preprocesado
    "scipy",
    "joblib",
]

# Modelos híbridos: combinan técnicas supervisadas y no supervisadas
# Casos de uso:
#   - PCA/UMAP → reducir dimensiones → clasificador
#   - KMeans → distancias a centroides como features → RF/LogReg
#   - IsolationForest → score de anomalía como feature → clasificador
#   - Semi-supervisado: LabelPropagation / LabelSpreading
hibrido = [
    "scikit-learn",   # PCA, KMeans, IsolationForest, pipelines, semi-supervisado
    "scipy",          # estadísticos, tests de normalidad
    "umap-learn",     # UMAP como alternativa a PCA (no lineal)
    "lightgbm",       # clasificador final potente
    "xgboost",        # clasificador final alternativo
    "imbalanced-learn",
    "joblib",
]

[tool.setuptools]
packages = ["{{ project_slug }}"]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"
markers   = [
    "smoke: test de humo — verifica que el pipeline arranca sin errores con datos sintéticos",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
